+++ src/index.ts
@@ -6,3 +6,4 @@ export { ResponseError, FetchError, RequiredError } from "./runtime";
 export type * from "./resources";
 export type * from "./domain";
 export * from "./webhook-events";
+export { CreateOrderRequestPrescriptionsInnerClinicalCompoundingReasonCategoryEnum as CompoundingReason } from "./models/CreateOrderRequestPrescriptionsInnerClinicalCompoundingReason";